extra effects on die sides should work basically the same as non basic effects.
I don't recommend trying to implement effects as resources though, since you would need to make a gd file for each effect anyways, and if you're gonna do that, then making a .tres off of a status effect just so you can drag it over to the inspector, when there's nothing you would be able to adjust for it in the inspector. The only thing that would do is make a bunch of unnecessary .tres files. A much more elegant way is to export an array of enums, where each enum value corresponds to a status effect.