added icons since it's convenient.
note: not gonna mess much with programming and ui related issues on this branch. Currently the way focus, disabled, etc works looks really weird with the new icons, but I don't wanna fix it yet. In any case, a bunch of that stuff may get dealt with when Aidan makes the map generation yeet these icons.