Changed the boss to be its own encounter
The way Ivan did it changed every corrupt die mage to be the boss. I made the boss its own thing using the sprite that was in progress. Unfortunately the sprite is so big it shifts the screen and causes a visual bug that I dont know how to fix